新增用戶用餐紀錄分頁及排序功能

+++ src/main/java/com/chumore/ordermaster/dto/RestDiningDto.java
@@ -1,6 +1,5 @@
 package com.chumore.ordermaster.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
@@ -12,7 +11,7 @@ public class RestDiningDto {
 	private String servedDatetime;
 	private String tableNumber;
 	private String memberName;
-	private BigDecimal totalPrice;
+	private Double totalPrice;
 	
 	public RestDiningDto() {
 		super();
@@ -67,10 +66,10 @@ public class RestDiningDto {
 		this.memberName = memberName;
 	}
 
-	public BigDecimal getTotalPrice() {
+	public Double getTotalPrice() {
 		return totalPrice;
 	}
-	public void setTotalPrice(BigDecimal totalPrice) {
+	public void setTotalPrice(Double totalPrice) {
 		this.totalPrice = totalPrice;
 	}
